moved datatypes shared between page and liveview to make them visible to both
